feat: Enhance OptimizedRelationshipInference with type hints and caching improvements
- Updated type hints for _compiled_verb_patterns and _type_inference_rules in optimized_relationship_inference.py.
- Added type hints for relationships in the extract_relationships method.
- Improved the benchmark_optimization function with a return type hint.

feat: Add optimizer_pipeline field to PipelineJSONLogger

- Included "optimizer_pipeline" field in the payload for logging events in pipeline_json_logger.py.

feat: Introduce new TypedDicts for validated query parameters and fallback query plans

- Added ValidatedQueryParameters, FallbackQueryPlan, and EnhancedTraversalParameters to query_optimizer_types.py.

fix: Improve error handling in schema_validator.py

- Moved OntologyValidationError class definition under TYPE_CHECKING to avoid import errors during runtime.
- Enhanced OntologySchemaError to include error details.

feat: Update structured logging in logic theorem optimizers

- Added "optimizer_pipeline" field to payloads in logic_optimizer.py and unified_optimizer.py.

test: Enhance unit tests for query optimizer typed returns

- Updated tests to verify the presence of expected keys and structure in execution plans and results.

test: Improve semantic deduplication integration tests

- Modified test cases to use more robust entity pairs for semantic similarity checks.

chore: Add comprehensive benchmark for ontology extraction

- Created bench_ontology_extraction_baseline.py to benchmark various extraction strategies and domains.
- Generated JSON and markdown reports for benchmark results.

docs: Add contributing guidelines for optimizers module

- Created CONTRIBUTING.md to outline contribution standards and best practices for the optimizers module.

+++ ipfs_datasets_py/optimizers/graphrag/ontology_comparator.py
@@ -94,9 +94,9 @@ class OntologyComparator:
         baseline_overall = getattr(baseline_score, 'overall', 0)
         
         if baseline_overall == 0:
-            improvement_pct = 0
+            improvement_pct = 0.0
         else:
-            improvement_pct = ((target_overall - baseline_overall) / baseline_overall * 100)
+            improvement_pct = float((target_overall - baseline_overall) / baseline_overall * 100)
         
         return {
             'improvement_percent': improvement_pct,
@@ -194,11 +194,11 @@ class OntologyComparator:
     # Custom Scoring
     def reweight_score(self, score: Any, weights: Dict[str, float]) -> float:
         """Reweight score with custom weights."""
-        weighted_sum = 0
-        total_weight = 0
+        weighted_sum = 0.0
+        total_weight = 0.0
         for dim, weight in weights.items():
             if weight > 0:
-                val = getattr(score, dim, 0)
+                val = float(getattr(score, dim, 0.0))
                 weighted_sum += val * weight
                 total_weight += weight
